Moved autenticate.ts to utils folder
