+++ api/core/polls.ts
@@ -1,10 +1,10 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node'
 import { db } from '../../config/firebaseConfig'
 import { Timestamp } from 'firebase-admin/firestore'
-import authenticate from '../../config/authenticate'
-import { validate } from '../../utils/validation'
-import { pollCreateDTO } from '../../models/schemas'
 import type Poll from '../../models/dataInterfaces'
+import { pollCreateDTO } from '../../models/schemas'
+import { validate } from '../../utils/validation'
+import authenticate from '../../utils/authenticate'
 
 export default async function handler(req: VercelRequest, res: VercelResponse) {
   // Apply authentication middleware
